+++ create-vtj/templates/h5/src/App.vue
@@ -1,0 +1,14 @@
+<template>
+  <ConfigProvider>
+    <Suspense>
+      <RouterView :key="route.fullPath"></RouterView>
+    </Suspense>
+  </ConfigProvider>
+</template>
+<script setup lang="ts">
+  // @ts-ignore
+  import { Suspense } from 'vue';
+  import { RouterView, useRoute } from 'vue-router';
+  import { ConfigProvider } from 'vant';
+  const route = useRoute();
+</script>