+++ packages/designer/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.20](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.9.19...@vtj/designer@0.9.20) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.9.19](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.9.18...@vtj/designer@0.9.19) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.9.18](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.9.17...@vtj/designer@0.9.18) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.9.17](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.9.16...@vtj/designer@0.9.17) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.9.16](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.9.15...@vtj/designer@0.9.16) (2025-01-13)
+
+
+### Features
+
+* ✨ h5平台 ([0497ccb](https://gitee.com/newgateway/vtj/commits/0497ccbb53f01a537176cae36f69b5d3019fc68e))
+* ✨ H5平台协议 ([51484a8](https://gitee.com/newgateway/vtj/commits/51484a8d0723897e462e1e2c37e16dbdf2546bea))
+
+
+
+
+
 ## [0.9.15](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.9.14...@vtj/designer@0.9.15) (2025-01-13)