+++ packages/materials/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.20](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.9.19...@vtj/materials@0.9.20) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.9.19](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.9.18...@vtj/materials@0.9.19) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.9.18](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.9.17...@vtj/materials@0.9.18) (2025-01-14)
+
+
+### Bug Fixes
+
+* 🐛 vant物料名称错误 ([07ad042](https://gitee.com/newgateway/vtj/commits/07ad0423c89303a744935627f9e405d317ea5f5e))
+
+
+
+
+
+## [0.9.17](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.9.16...@vtj/materials@0.9.17) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.9.16](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.9.15...@vtj/materials@0.9.16) (2025-01-13)
+
+
+### Features
+
+* ✨ h5平台 ([0497ccb](https://gitee.com/newgateway/vtj/commits/0497ccbb53f01a537176cae36f69b5d3019fc68e))
+
+
+
+
+
 ## [0.9.15](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.9.14...@vtj/materials@0.9.15) (2025-01-13)
 
 **Note:** Version bump only for package @vtj/materials